requirements.txt
django==2.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
djangorestframework-camel-case==0.2.0
drf-dynamic-fields==0.3.0
pyjwt==1.6.4
psycopg2==2.7.5
